+++ spec/hitnmiss/driver_spec.rb
@@ -74,16 +74,7 @@ describe Hitnmiss::Driver do
     describe "#set" do
       it "raises error indicating not implemented" do
         expect { subject.set(double('key'),
-                             double('value'),
-                             double('expiration')) }.to \
-        raise_error(Hitnmiss::Errors::NotImplemented)
-      end
-    end
-
-    describe "#setnoexp" do
-      it "raises error indicating not implemented" do
-        expect { subject.setnoexp(double('key'),
-                                  double('value')) }.to \
+                             double('entity')) }.to \
         raise_error(Hitnmiss::Errors::NotImplemented)
       end
     end